+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- 🎉 feat: added DropDownMenu component
 - 🎉 feat: dropdown added
 - 🎉 feat: added responsiveness to everything that didn't + scroll snap added
 - 🎉 feat: added two profile cards + one shop card